更新版本

+++ game-api/DockerBuild.bat
@@ -1,9 +1,7 @@
-
+@echo off
 rem must copy file...
-xcopy ..\game-api-scripts\src game-api-scripts\src\ /s /e
+xcopy ..\game-api-scripts\src game-api-scripts\src\ /s /e /q 2>nul
 
 docker image build -t game-api:releases .
 
-rd /s/q game-api-scripts
-
-
+if exist game-api-scripts rd /s/q game-api-scripts
