+++ game-hall/DockerBuild.bat
@@ -1,9 +1,7 @@
-
+@echo off
 rem must copy file...
-xcopy ..\game-hall-scripts\src game-hall-scripts\src\ /s /e
+xcopy ..\game-hall-scripts\src game-hall-scripts\src\ /s /e /q 2>nul
 
 docker image build -t game-hall:releases .
 
-rd /s/q game-hall-scripts
-
-
+if exist game-hall-scripts rd /s/q game-hall-scripts
